+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-01 15:47:15
+
+### Changed Files
+- `src/features/policies/hooks/usePolicyForm.ts`
+
+### Statistics
+```
+ 1 file changed, 11 insertions(+), 2 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-01 15:37:09
 
 ### Changed Files